Fixes and additional Assets
Added Inverse Kinematics to every single Animation (Pro Tipp: do it up front, it's a shitload of work and kills everythign you had done so far)
Added ClimbingDaggers, a bunch of experimental Assets fo rthe fun of it.
Started with the statues for the myth
Added the Hydra as the first Boss.
Corrected Chests and pieces of environment